feat: memory info widget implemented

+++ SystemMonitor/include/CustomChartView.h
@@ -12,7 +12,7 @@
 class CustomChartView : public QChartView
 {
 public:
-	CustomChartView(QWidget* parent = nullptr, int minPointsX = 0, int maxPointsX = 1, QString titleX = "", int minPointsY = 0, int maxPointsY = 1, QString titleY = "", QColor borderColor1 = QColor(150, 150, 250, 255), QColor fillColor1 = QColor(150, 150, 250, 100), QColor borderColor2 = QColor(230, 122, 50, 255), QColor fillColor2 = QColor(230, 122, 50, 100));
+	CustomChartView(QWidget* parent = nullptr, double minPointsX = 0, double maxPointsX = 1, QString titleX = "", double minPointsY = 0, double maxPointsY = 1, QString titleY = "", QColor borderColor1 = QColor(150, 150, 250, 255), QColor fillColor1 = QColor(150, 150, 250, 100), QColor borderColor2 = QColor(230, 122, 50, 255), QColor fillColor2 = QColor(230, 122, 50, 100));
 
 	void append(double data);
 	void append(double data1, double data2);
@@ -31,8 +31,8 @@ private:
 	QLineSeries* m_lowerSeries2;
 	QAreaSeries* m_areaSeries2;
 
-	int m_maxPointsX;
-	int m_maxPointsY;
+	double m_maxPointsX;
+	double m_maxPointsY;
 	std::deque<double> m_dataPoints1;
 	std::deque<double> m_dataPoints2;
